dashappv2/requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudstorage==0.10.0
dash==1.7.0
dash-bootstrap-components==0.7.2
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
decorator==4.4.1
ez-setup==0.9
Flask==1.1.1
Flask-Compress==1.4.0
fsspec==0.6.1
future==0.18.2
gcloud==0.18.3
gcsfs==0.5.3
gitdb2==2.0.6
GitPython==3.0.5
google-api-core==1.14.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.25.3
plotly==4.3.0
protobuf==3.11.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-git==2018.2.1
python-magic==0.4.15
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scipy==1.3.1
Send2Trash==1.5.0
simpy==3.0.11
simpy.io==0.2.3
six==1.13.0
smmap2==2.0.5
urllib3==1.25.7
webapp2==2.5.2
webapp3==1.0.2
WebOb==1.8.5
Werkzeug==0.16.0
wincertstore==0.2